requirements.txt
eth-brownie
pytest